move WithLongRunningRequestTermination to generic filters
