+++ staging/src/k8s.io/apiserver/pkg/server/filters/longrunning.go
@@ -17,10 +17,13 @@ limitations under the License.
 package filters
 
 import (
+	"context"
 	"net/http"
 	"strings"
+	"time"
 
 	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/apiserver/pkg/endpoints/request"
 	apirequest "k8s.io/apiserver/pkg/endpoints/request"
 )
 
@@ -38,4 +41,84 @@ func BasicLongRunningRequestCheck(longRunningVerbs, longRunningSubresources sets
 		}
 		return false
 	}
+}
+
+// WithLongRunningRequestTermination starts closing long running requests upon receiving TerminationStartCh signal.
+//
+// It does it by running two additional go routines.
+// One for running the request and the second one for intercepting the termination signal and propagating it to the requests.
+// If the request is not terminated within 5 seconds it will be forcefully killed.
+//
+// This filter exists because sometimes propagating the termination signal is not enough.
+// It turned out that long running requests might block on:
+//  io.Read() for example in https://golang.org/src/net/http/httputil/reverseproxy.go
+//  http2.(*serverConn).writeDataFromHandler which might be actually an issue with the std lib itself
+//
+// Instead of trying to identify current and future issues we provide a filter that ensures terminating long running requests.
+//
+// Also note that upon receiving termination signal the http server sends
+// sends GOAWAY with ErrCodeNo to tell the client we're gracefully shutting down.
+// But the connection isn't closed until all current streams are done.
+func WithLongRunningRequestTermination(handler http.Handler, longRunningFunc apirequest.LongRunningRequestCheck, terminationStartCh <-chan struct{}) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		requestInfo, found := request.RequestInfoFrom(req.Context())
+		if !found || longRunningFunc == nil {
+			handler.ServeHTTP(w, req)
+			return
+		}
+
+		isLongRunning := longRunningFunc(req, requestInfo)
+		if !isLongRunning {
+			handler.ServeHTTP(w, req)
+			return
+		}
+
+		ctx, cancelCtxFn := context.WithCancel(req.Context())
+		defer cancelCtxFn()
+		req = req.WithContext(ctx)
+
+		errCh := make(chan interface{}, 2)
+		// note that is is okay to leave the errCh open
+		// eventually it will be garbage collected
+		doneCh := make(chan struct{})
+
+		go func() {
+			defer func() {
+				err := recover()
+				select {
+				case errCh <- err:
+					return
+				case <-doneCh:
+					return
+				}
+			}()
+			select {
+			case <-terminationStartCh:
+				cancelCtxFn()
+				time.Sleep(5 * time.Second)
+				panic(http.ErrAbortHandler)
+			case <-doneCh:
+				return
+			}
+		}()
+
+		go func() {
+			defer func() {
+				err := recover()
+				select {
+				case errCh <- err:
+					return
+				case <-doneCh:
+					return
+				}
+			}()
+			handler.ServeHTTP(w, req)
+		}()
+
+		err := <-errCh
+		close(doneCh)
+		if err != nil {
+			panic(err)
+		}
+	})
 }
